[reporters] Move FontbakeryReporter base class to module file.
