+++ Lib/fontbakery/reporters/terminal.py
@@ -30,6 +30,7 @@ except ImportError:
   # Python 3
   import builtins as __builtin__
 
+from fontbakery.reporters import FontbakeryReporter
 
 from fontbakery.testrunner import (
               DEBUG
@@ -45,7 +46,6 @@ from fontbakery.testrunner import (
             , ENDSECTION
             , START
             , END
-            , ProtocolViolationError
             )
 
 statuses = (
@@ -120,87 +120,6 @@ UNICORN = r"""
               No test is failing.
     <<Art by Colin J. Randall, cjr, 10mar02>>
 """
-
-class FontbakeryReporter(object):
-  def __init__(self, is_async=False, runner=None):
-    self._started = None
-    self._ended = None
-    self._order = None
-    self._results = [] # ENDTEST events in order of appearance
-    self._indexes = {}
-    self._tick = 0
-    self._counter = Counter()
-
-    # Runner should know if it is async!
-    self.is_async = is_async
-    self.runner = runner
-
-  def run(self):
-    """
-    self.runner must be present
-    """
-    for event in self.runner.run():
-      self.receive(event)
-
-  @property
-  def order(self):
-    return self._order
-
-  def _get_key(self, identity):
-    section, test, iterargs = identity
-    return ('{}'.format(section) if section else section
-          , '{}'.format(test) if test else test
-          , iterargs
-          )
-
-  def _get_index(self, identity):
-    key = self._get_key(identity)
-    try:
-      return self._indexes[key]
-    except KeyError:
-      self._indexes[key] = len(self._indexes)
-      return self._indexes[key]
-
-  def _set_order(self, order):
-    self._order = order
-    length = len(order)
-    self._counter['(not finished)'] = length - len(self._results)
-    self._indexes = dict(zip(map(self._get_key, order), range(length)))
-
-  def _cleanup(self, (status, message, identity)):
-    pass
-
-  def _output(self, (status, message, identity)):
-    pass
-
-  def _register(self, event):
-    status, message, identity = event
-    self._tick += 1
-    if status == START:
-      self._set_order(message)
-      self._started = event
-
-    if status == END:
-      self._ended = event
-
-    if status == ENDTEST:
-      self._results.append(event)
-      self._counter[message.name] += 1
-      self._counter['(not finished)'] -= 1
-
-  def receive(self, event):
-    status, message, identity = event
-    if self._started is None and status != START:
-      raise ProtocolViolationError('Received Event before status START: '\
-                                      ' {} {}.'.format(status, message))
-    if self._ended:
-      status, message, identity = event
-      raise ProtocolViolationError('Received Event after status END: '\
-                                        '{} {}.'.format(status, message))
-    self._register(event)
-    self._cleanup(event)
-    self._output(event)
-
 
 class TerminalProgress(FontbakeryReporter):
   def __init__(self, print_progress=True
